[Fix] #396 - ProfileImageView 공통 컴포넌트로 변경
