+++ SOPT-iOS/Projects/Features/PokeFeature/Interface/Sources/Enum/PokeRelation.swift
@@ -8,6 +8,10 @@
 
 import Foundation
 
+public protocol ProfileRelation {
+    
+}
+
 public enum PokeRelation: String {
     case nonFriend
     case newFriend = "친한친구"
